+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@
 - Changes:
   - Renamed `from_hexip` to `from_hexipv4` and created an aliases `from_hexip` for `from_hexipv4`. Same for `to_hexipv4` and bang methods.
 - New methods: `from_hexipv6` and `from_hexipv6!`
+- Fix:
+  - bin2hex: fix odd byte cropping issue
 
 ## [2.3.0]